+++ source/BattleMiniMap/src/Config/BattleMiniMapOptionClassFactory.cs
@@ -1,5 +1,4 @@
-﻿using BattleMiniMap.View.MapTerrain;
-using MissionLibrary.Provider;
+﻿using MissionLibrary.Provider;
 using MissionLibrary.View;
 using MissionSharedLibrary.Provider;
 using MissionSharedLibrary.View.ViewModelCollection;
@@ -7,7 +6,6 @@ using MissionSharedLibrary.View.ViewModelCollection.Options;
 using MissionSharedLibrary.View.ViewModelCollection.Options.Selection;
 using TaleWorlds.Core;
 using TaleWorlds.GauntletUI;
-using TaleWorlds.MountAndBlade;
 
 namespace BattleMiniMap.Config
 {