+++ source/BattleMiniMap/src/Config/HotKey/GameKeyConfig.cs
@@ -1,6 +1,6 @@
-﻿using System.IO;
-using MissionSharedLibrary.Config;
+﻿using MissionSharedLibrary.Config;
 using MissionSharedLibrary.Config.HotKey;
+using System.IO;
 
 namespace BattleMiniMap.Config.HotKey
 {