+++ source/BattleMiniMap/src/View/AgentMarkers/Colors/AgentMarkerColorType.cs
@@ -1,7 +1,6 @@
 ﻿using TaleWorlds.Library;
-using TaleWorlds.MountAndBlade;
 
-namespace BattleMiniMap.View.AgentMarker.Colors
+namespace BattleMiniMap.View.AgentMarkers.Colors
 {
     public enum AgentMarkerColorType
     {
@@ -15,7 +14,7 @@ namespace BattleMiniMap.View.AgentMarker.Colors
             switch (colorType)
             {
                 case AgentMarkerColorType.Inactive:
-                    return new Color(0.3f, 0.3f, 0.3f, 0.8f);
+                    return new Color(0.5f, 0.4f, 0.4f, 0.8f);
                 case AgentMarkerColorType.PlayerTeam:
                     return new Color(0.25f, 0.60f, 0.94f);
                 case AgentMarkerColorType.PlayerAlly:
@@ -23,9 +22,9 @@ namespace BattleMiniMap.View.AgentMarker.Colors
                 case AgentMarkerColorType.PlayerEnemy:
                     return new Color(0.91f, 0.18f, 0.32f);
                 case AgentMarkerColorType.Human:
-                    return new Color(0.6f, 0.6f, 0.3f);
+                    return new Color(0.8f, 0.8f, 0.2f);
                 case AgentMarkerColorType.Horse:
-                    return new Color(0.5f, 0.5f, 0.5f);
+                    return new Color(0.3f, 0.3f, 0.2f);
                 default:
                     return new Color(0.4f, 0.4f, 0.2f);
             }