+++ source/BattleMiniMap/src/View/AgentMarkers/TextureProviders/HorseAgentTextureProvider.cs
@@ -1,13 +1,20 @@
-﻿using System.Drawing;
+﻿using BattleMiniMap.View.Image;
+using System.Drawing;
 using System.Drawing.Imaging;
-using BattleMiniMap.View.Image;
 using TaleWorlds.TwoDimension;
 
-namespace BattleMiniMap.View.AgentMarker.TextureProviders
+namespace BattleMiniMap.View.AgentMarkers.TextureProviders
 {
     class HorseAgentTextureProvider : IAgentTextureProvider
     {
+        private Bitmap _bitmap;
         private Texture _texture;
+
+        public Bitmap GetBitmap()
+        {
+            return _bitmap ??= CreateBitMap();
+        }
+
         public Texture GetTexture()
         {
             return _texture ??= CreateTexture();
@@ -32,12 +39,17 @@ namespace BattleMiniMap.View.AgentMarker.TextureProviders
         //    });
         //}
 
-        private Texture CreateTexture()
+        private Bitmap CreateBitMap()
         {
             var bitmap = new Bitmap(50, 50, PixelFormat.Format32bppArgb);
             var graphics = Graphics.FromImage(bitmap);
             Fill(graphics, new SolidBrush(Color.FromArgb(255, 255, 255, 255)));
-            return bitmap.CreateTexture();
+            return bitmap;
+        }
+
+        private Texture CreateTexture()
+        {
+            return GetBitmap().CreateTexture();
         }
 
         private void Fill(Graphics graphics, Brush brush)