+++ source/BattleMiniMap/src/View/AgentMarkers/TextureProviders/HumanAgentTextureProvider.cs
@@ -3,23 +3,35 @@ using System.Drawing;
 using System.Drawing.Imaging;
 using TaleWorlds.TwoDimension;
 
-namespace BattleMiniMap.View.AgentMarker.TextureProviders
+namespace BattleMiniMap.View.AgentMarkers.TextureProviders
 {
     public class HumanAgentTextureProvider : IAgentTextureProvider
     {
+        private Bitmap _bitmap;
         private Texture _texture;
+
+        public Bitmap GetBitmap()
+        {
+            return _bitmap ??= CreateBitMap();
+        }
+
         public Texture GetTexture()
         {
             return _texture ??= CreateTexture();
         }
 
-        private Texture CreateTexture()
+        private Bitmap CreateBitMap()
         {
             var bitmap = new Bitmap(50, 50, PixelFormat.Format32bppArgb);
             var graphics = Graphics.FromImage(bitmap);
             //Draw(graphics, new Pen(Color.FromArgb(255, 50, 50, 50), 3));
             Fill(graphics, new SolidBrush(Color.FromArgb(255, 255, 255, 255)));
-            return bitmap.CreateTexture();
+            return bitmap;
+        }
+
+        private Texture CreateTexture()
+        {
+            return GetBitmap().CreateTexture();
         }
 
         //private void Draw(Graphics graphics, Pen pen)