+++ source/BattleMiniMap/src/View/AgentMarkers/TextureProviders/IAgentTextureProvider.cs
@@ -1,0 +1,11 @@
+﻿using System.Drawing;
+using TaleWorlds.TwoDimension;
+
+namespace BattleMiniMap.View.AgentMarkers.TextureProviders
+{
+    public interface IAgentTextureProvider
+    {
+        Bitmap GetBitmap();
+        Texture GetTexture();
+    }
+}